Requisition Full Functionality

+++ ERP/src/main/java/com/ERP/dao/ProjectDao.java
@@ -7,7 +7,7 @@ import com.ERP.model.Project;
 
 public interface ProjectDao {
 
-	Project findByProjectid(int projectId);
+	Project findByProjectid(long projectId);
 	
 	Project findByProjectname(String projectName);
